refactor: dissolve triggers/ipfs.ts
